Begin moving driver interaction code to another file
