+++ pkg/ebpf/tracer_windows.go
@@ -2,20 +2,10 @@
 
 package ebpf
 
-/*
-//! Defines the objects used to communicate with the driver as well as its control codes
-#include "c/ddfilterapi.h"
-
-//! These includes are needed to use constants defined in the ddfilterapi
-#include <WinDef.h>
-#include <WinIoCtl.h>
-*/
-import "C"
 import (
-	"encoding/binary"
+	"C"
 	"expvar"
 	"fmt"
-	"net"
 	"sync/atomic"
 	"time"
 	"unsafe"
@@ -23,13 +13,10 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/process/util"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
-	"golang.org/x/net/ipv4"
 	"golang.org/x/sys/windows"
 )
 
 const (
-	driverFile = `\\.\ddfilter`
-
 	// Number of buffers to use with the IOCompletion port to communicate with the driver
 	totalReadBuffers = 32
 )
@@ -37,19 +24,7 @@ const (
 var (
 	expvarEndpoints map[string]*expvar.Map
 	expvarTypes     = []string{"driver_total_stats", "driver_handle_stats", "packet_count"}
-
-	// Buffer holding datadog driver filterapi (ddfilterapi) signature to ensure consistency with driver.
-	ddAPIVersionBuf = makeDDAPIVersionBuffer(C.DD_FILTER_SIGNATURE)
 )
-
-// Creates a buffer that Driver will use to verify proper versions are communicating
-// We create a buffer because the system calls we make need a *byte which is not
-// possible with const value
-func makeDDAPIVersionBuffer(signature uint64) []byte {
-	buf := make([]byte, C.sizeof_uint64_t)
-	binary.LittleEndian.PutUint64(buf, signature)
-	return buf
-}
 
 func init() {
 	expvarEndpoints = make(map[string]*expvar.Map, len(expvarTypes))
@@ -60,44 +35,28 @@ func init() {
 
 // Tracer struct for tracking network state and connections
 type Tracer struct {
-	config       *Config
-	driverHandle windows.Handle
-	iocp         windows.Handle
-	bufs         []readBuffer
-	packetCount  int64
+	config           *Config
+	driverController *DriverController
+	bufs             []readBuffer
+	packetCount      int64
 }
 
 // NewTracer returns an initialized tracer struct
 func NewTracer(config *Config) (*Tracer, error) {
-	handle, err := openDriverFile(driverFile)
+	dc, err := NewDriverController()
 	if err != nil {
-		return nil, fmt.Errorf("%s : %s", "could not create driver handle", err)
-	}
-
-	// Create IO Completion port that we'll use to communicate with the driver
-	iocp, err := windows.CreateIoCompletionPort(handle, windows.Handle(0), 0, 0)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create IO completion port %v", err)
+		return nil, fmt.Errorf("could not create windows driver controller", err)
 	}
 
 	tr := &Tracer{
-		driverHandle: handle,
-		iocp:         iocp,
-		bufs:         make([]readBuffer, totalReadBuffers),
+		driverController: dc,
+		bufs:             make([]readBuffer, totalReadBuffers),
 	}
 
-	// Set the packet filters that will determine what we pull from the driver
-	// TODO: Determine failure condition for not setting filter
-	// TODO: I.e., one or more, all fail? I.E., at what point do we not create a tracer
-	err = tr.prepareDriverFilters()
+	// We want tracer to own the buffers, but the DriverController to make the calls to set them up
+	tr.bufs, err = dc.prepareReadBuffers(tr.bufs)
 	if err != nil {
-		return nil, fmt.Errorf("failed to setup packet filters on the driver: %v", err)
-	}
-
-	// Prepare the read buffers that will be used to send packets from kernel to user space
-	err = tr.prepareReadBuffers()
-	if err != nil {
-		return nil, fmt.Errorf("failed to prepare ReadBuffers: %v", err)
+		return nil, fmt.Errorf("error preparing driver's read buffers")
 	}
 
 	err = tr.initPacketPolling()
@@ -113,6 +72,7 @@ func (t *Tracer) Stop() {}
 
 func (t *Tracer) expvarStats() {
 	ticker := time.NewTicker(5 * time.Second)
+	defer ticker.Stop()
 	// starts running the body immediately instead waiting for the first tick
 	for range ticker.C {
 		stats, err := t.GetStats()
@@ -218,82 +178,6 @@ func (t *Tracer) GetStats() (map[string]interface{}, error) {
 	}, nil
 }
 
-// NewDDAPIFilter returns a filter we can apply to the driver
-func newDDAPIFilter(direction C.uint64_t, ifaceIndex int, isIPV4 bool) (fd C.struct__filterDefinition) {
-	fd.filterVersion = C.DD_FILTER_SIGNATURE
-	fd.size = C.sizeof_struct__filterDefinition
-	fd.direction = direction
-
-	if isIPV4 {
-		fd.af = windows.AF_INET
-		fd.v4InterfaceIndex = (C.ulonglong)(ifaceIndex)
-	} else {
-		fd.af = windows.AF_INET6
-		fd.v6InterfaceIndex = (C.ulonglong)(ifaceIndex)
-	}
-
-	return fd
-}
-
-func (t *Tracer) setFilter(fd C.struct__filterDefinition) (err error) {
-	var id int64
-	err = windows.DeviceIoControl(t.driverHandle, C.DDFILTER_IOCTL_SET_FILTER, (*byte)(unsafe.Pointer(&fd)), uint32(unsafe.Sizeof(fd)), (*byte)(unsafe.Pointer(&id)), uint32(unsafe.Sizeof(id)), nil, nil)
-	if err != nil {
-		return log.Errorf("Failed to set filter: %s\n", err.Error())
-	}
-	return
-}
-
-// To capture all traffic for an interface, we create an inbound/outbound traffic filter
-// for both IPV4 and IPV6 traffic going to that interface
-func createFiltersForInterface(iface net.Interface) (filters []C.struct__filterDefinition) {
-	filters = append(filters, newDDAPIFilter(C.DIRECTION_INBOUND, iface.Index, true))
-	filters = append(filters, newDDAPIFilter(C.DIRECTION_OUTBOUND, iface.Index, true))
-	filters = append(filters, newDDAPIFilter(C.DIRECTION_INBOUND, iface.Index, false))
-	filters = append(filters, newDDAPIFilter(C.DIRECTION_OUTBOUND, iface.Index, false))
-	return
-}
-
-func (t *Tracer) prepareDriverFilters() (err error) {
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		log.Errorf("Error getting interfaces: %s\n", err.Error())
-		return
-	}
-
-	for _, i := range ifaces {
-		log.Debugf("Setting filters for interface: %s [%+v]", i.Name, i)
-
-		for _, filter := range createFiltersForInterface(i) {
-			err = t.setFilter(filter)
-			if err != nil {
-				log.Warnf("Failed to set filter [%+v] on interface [%+v]: %s\n", filter, i, err.Error())
-			}
-		}
-	}
-	return nil
-}
-
-// Add buffers to Tracer object. Even though the windows API will actually
-// keep the buffers, add it to the tracer so that golang doesn't garbage collect
-// the buffers out from under us
-func (t *Tracer) prepareReadBuffers() (err error) {
-	for i := 0; i < totalReadBuffers; i++ {
-		err = windows.ReadFile(t.driverHandle, t.bufs[i].data[:], nil, &t.bufs[i].ol)
-		if err != nil {
-			if err != windows.ERROR_IO_PENDING {
-				fmt.Printf("failed to initiate readfile %v\n", err)
-				windows.CloseHandle(t.iocp)
-				t.iocp = windows.Handle(0)
-				t.bufs = nil
-				return
-			}
-		}
-	}
-	err = nil
-	return
-}
-
 // DebugNetworkState returns a map with the current tracer's internal state, for debugging
 func (t *Tracer) DebugNetworkState(clientID string) (map[string]interface{}, error) {
 	return nil, ErrNotImplemented
@@ -307,43 +191,4 @@ func (t *Tracer) DebugNetworkMaps() (*Connections, error) {
 // CurrentKernelVersion is not implemented on this OS for Tracer
 func CurrentKernelVersion() (uint32, error) {
 	return 0, ErrNotImplemented
-}
-
-// readBuffer is the buffer to pass into ReadFile system call to pull out packets
-type readBuffer struct {
-	ol   windows.Overlapped
-	data [128]byte
-}
-
-func (rb *readBuffer) printPacket() {
-	var header ipv4.Header
-	var pheader C.struct_filterPacketHeader
-	dataStart := unsafe.Sizeof(pheader)
-	log.Infof("data start is %d\n", dataStart)
-	header.Parse(rb.data[dataStart:])
-	log.Infof(" %v    ==>>    %v", header.Src.String(), header.Dst.String())
-}
-
-func openDriverFile(path string) (windows.Handle, error) {
-	p, err := windows.UTF16PtrFromString(path)
-	if err != nil {
-		return windows.InvalidHandle, err
-	}
-	log.Debug("Creating Driver handle...")
-	h, err := windows.CreateFile(p,
-		windows.GENERIC_READ|windows.GENERIC_WRITE,
-		windows.FILE_SHARE_READ|windows.FILE_SHARE_WRITE,
-		nil,
-		windows.OPEN_EXISTING,
-		windows.FILE_FLAG_OVERLAPPED,
-		windows.Handle(0))
-	if err != nil {
-		return windows.InvalidHandle, err
-	}
-	log.Info("Connected to driver and handle created")
-	return h, nil
-}
-
-func closeDriverFile(handle windows.Handle) error {
-	return windows.CloseHandle(handle)
 }
